Add Abstractive Health Lab implementation with services and environment setup
