--- labs/openmrs/AbstractiveHealth Lab/.env
@@ -1,3 +1,0 @@
-AH_EMAIL=your_email_here
-AH_USERNAME=your_username_here
-AH_PASSWORD=your_password_here
